Change Feedpage to Suggestions and add to Feed
