+++ react-app/src/components/Suggestions/Suggestions.css
